Endpoint for Martin's authentication proxy
